# crates/api/src/openapi.yaml
openapi: 3.1.0
info:
  title: Tileforge API
  description: |
    REST API for processing images into XYZ map tiles.

    ## Authentication
    Most endpoints require a Bearer token (JWT) in the Authorization header.
    Pro users can generate API keys from the settings page.
  version: 0.1.0
  license:
    name: MIT
    url: https://github.com/thesandybridge/tileforge/blob/main/LICENSE
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.tileforge.dev
    description: Production
...
paths:
...
  /api/tilesets/{slug}/tiles.pmtiles:
    get:
      tags: [Tilesets]
      summary: Stream a hosted PMTiles archive
      description: Supports HTTP byte ranges. Public tilesets can be embedded directly; private tilesets require owner authentication.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: Range
          in: header
          schema:
            type: string
            example: bytes=0-16383
      responses:
        '200':
          description: Complete PMTiles archive
        '206':
          description: Requested PMTiles byte range
        '404':
          description: Tileset not found or inaccessible
...
  /api/tilesets:
    get:
      tags: [Tilesets]
      summary: List tilesets
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
        - name: user
          in: query
          description: Filter by username
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of tilesets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tileset'
    post:
      tags: [Tilesets]
      summary: Create tileset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, job_id]
              properties:
                name:
                  type: string
                job_id:
                  type: string
                  format: uuid
                public:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Tileset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tileset'
        '401':
          description: Unauthorized
        '409':
          description: Slug already exists
  /api/tilesets/{slug}:
    get:
      tags: [Tilesets]
      summary: Get tileset details
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tileset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tileset'
        '404':
          description: Tileset not found
    patch:
      tags: [Tilesets]
      summary: Update tileset
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                public:
                  type: boolean
      responses:
        '200':
          description: Tileset updated
        '401':
          description: Unauthorized
        '403':
          description: Not owner
        '404':
          description: Tileset not found
    delete:
      tags: [Tilesets]
      summary: Delete tileset
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Tileset deleted
        '401':
          description: Unauthorized
        '403':
          description: Not owner
        '404':
          description: Tileset not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: apiKey
      in: header
      name: X-Admin-Secret
  schemas:
    Tileset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        owner_id:
          type: string
          format: uuid
        owner_username:
          type: string
        job_id:
          type: string
          format: uuid
        public:
          type: boolean
        tile_count:
          type: integer
        min_zoom:
          type: integer
        max_zoom:
          type: integer
        source_epsg:
          type: [integer, "null"]
          description: EPSG code detected from a GeoTIFF source
        source_bounds:
          type: [array, "null"]
          description: Source bounds ordered west, south, east, north
          minItems: 4
          maxItems: 4
          items:
            type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...